Called storage logic from the correct module (MssqlEcto rather than MssqlEcto.Connection)
